Time: 58 ms (73.76%), Space: 61.7 MB (11.7%) - LeetHub

+++ 0011-container-with-most-water/0011-container-with-most-water.cpp
@@ -16,6 +16,6 @@ public:
             sum = max(sum,ans);
 
         }
-        return sum;
+        return (int)sum;
     }
 };
